+++ charts/core/src/ChartContext/ChartContext.types.ts
@@ -1,0 +1,14 @@
+import { ChartOptions, SeriesOption } from '../Chart/Chart.types';
+
+export interface ChartProviderProps {
+  chartOptions: any;
+  updateChartOptions: (newOptions: Partial<ChartOptions>) => void;
+  addChartSeries: (series: SeriesOption) => void;
+  removeChartSeries: (name: string) => void;
+}
+
+export interface ChartContextType {
+  chartOptions: any;
+  updateChartOptions: (newOptions: Partial<ChartOptions>) => void;
+  addChartSeries: (series: SeriesOption) => void;
+}
