--- charts/core/src/ChartContext/ChartProvider/index.ts
@@ -1,1 +1,0 @@
-export { ChartProvider } from './ChartProvider';
