--- charts/core/src/ChartContext/useChartContext/index.ts
@@ -1,1 +1,0 @@
-export { useChartContext } from './useChartContext';
